Second Project

+++ Day2.py
@@ -1,6 +1,6 @@
 # Second project
 print("Hello! Welcome to the tip calculator!")
-bill = int(input("Whats the total bill?: "))
+bill = float(input("Whats the total bill?: "))
 tip = int(input("How much tip would you like to give?: "))
 people = int(input("How many people to split the bill?: "))
